pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "innovati"
version = "0.1.0"
description = "Innovati Fullstack Technical Test"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]",
    "pydantic>=2",
    "SQLAlchemy>=2",
    "aiosqlite; python_version >= '3.11'",
    "asyncpg; python_version >= '3.11'",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
]

[project.optional-dependencies]
llm = [
    "openai>=1.40",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"